osd_client: refactor strings for ceph_hobject_id
Currently there is a mess with strings. In this patch I try
to avoid raw pointers and lengths and replace ->oid_name
with ceph_object_id, ->nspace and ->key raw pointers with
with ceph_string.

Why? In future ceph_hobject_id will be used on OSD server
side, so do not want to do strdup on hoid copy, which will
be implemented later.  Easier just to increase a ref on
ceph_string.