osd_client: make oloc_decode() full-fledged and pulish as an API
oloc_decode() does not actually parse ->pool_ns, but instead checked
if it has been changed or not.  Publish oloc_decode() as an API
ceph_oloc_decode() call and make it full-fledged, i.e. now it parses
->pool_ns and replaces the old string with the new one.

But careful, osd_client.c uses oloc_decode() specifically, and expects
that ->pool_ns is not changed, thus we have to increase a reference
and put it afterwards. Also check if ->pool_ns is changed or not moved
to the caller of oloc_decode().