messenger: switch data cursor from page to iov_iter
The first problem is performance. Why not to pass to read/write
socket function the whole iov_iter?  So better to make data cursor
as iov_iter, which is generic for many API calls.

The second reason is the support of kvec in the future, i.e. we
do not have a page in a hand, but a buffer.

So this page is a preparation, the iteration #1: users of data
cursor do not see pages, but use cursor->iter instead.  Internally
cursor still uses page.  In next patches that will be avoided.