sched: introduce preempt_count and preempt_schedule()
schedule() can be invoked only when preemption is enabled,
i.e. preempt_count == 0.  In order to be sure schedule()
is not called under the spinlock or rwlock this patch
introduces preempt counter and preempt_schedule() call,
which does schedule() when preemption is enabled.

Also in order to give the waker more freedom it makes sense
to do high priority wakeup, when we can't just put a new
task to the list, but we need to call schedule() immediately.
So new WF_HIGHPRI flag is introduced, and when it is used
task is put to the head, not to the tail, so next schedule()
pops the task and runs it reducing the latency.