osd_server: embed input data and cursor inside ceph_osd_req_op
In future patches all operations will become a part of a transaction,
thus we should be careful and do not allocate input data and cursor
on the stack. Now the osd operation is the self-sufficient structure,
which contains all members, which are allocated not on the stack.

Seems the time when ceph_osd_req_op structure will be split on
client and server is comming. For now keep as is, i.e. only server
side uses ->indata, ->outdata, ->incur, but for client side this
only increases memory footprint.