osd_server: use memstore and transaction logic
All logic which operates with objects and blocks now is the part of
the memstore, osd_server switched to it in the current patch.

Also, transaction structure is used to accumulate mutative operations,
which are executed locally inside a store or replicated on other
osds (in future patches).

What is not yet done is that transaction can fail and return code
should be set to the corresponding operation (not yet clear how
crimson or legacy osd deal with this.