osd_server,transaction,memstore: make handle_osd_op_stat part of the osd_server
This is not nice and I do not like it, but the fact is the following:
we need an object cache outside of the store to be able to operate
with object mtime and size in a single transaction. In simple words
we need object modifications to be seen in current transaction.

Would be great to have this particular logic in the store itself.
Need to think carefully about this.

Also, according to current implementation other transactions will
see the object is modified inside a cache. This can be fixed in the
future.