main: switch signal delivery from signalfd to eventfd
Why? In the next patch userpoll will be used, which it turned out to
be is not a best friend with signalfd. Why? The problem is that kernel
function signalfd_poll() checks pending signals for *current* process,
but userpoll offloads vfs_epoll() call to a worker (see
ep_poll_callback_work()), which means pending signals of the worker
will be checked and 0 is returned, obviously not what we expected.